// src/app/model/alerts.ts
export class Alerts {
    lp: number;
    instrument: string;
    lastPrice: string;
    bidPrice: string;
    oferPrice: string;
    percent: string;
    levelPrice: string;
    levelNo: string;
    levelPriceNext: string;
    tickDistance: number;
    name: string;
    tickSize: string;
    colorAlert: string;

    deliveryMonth: string;
    firstNotice: string;
    expiration: string;

}